# Improve benchmark capability. Use it to improve performance of MultiplyByPrice by over 2x