### Changelist
- Add small improvements to the `make benchmark` functionality
- make `multiplyByPrice` public for testing (now `MultiplyByPrice`)
- Improve multiply by price

### Test Plan

Tested locally on an M2 Macbook
```
Before:
BenchmarkMultiplyByPrice-8   	 2943225	       409.0 ns/op	     504 B/op	      18 allocs/op
BenchmarkMultiplyByPrice-8   	 2944702	       408.6 ns/op	     504 B/op	      18 allocs/op
BenchmarkMultiplyByPrice-8   	 2935136	       411.0 ns/op	     504 B/op	      18 allocs/op
BenchmarkMultiplyByPrice-8   	 2938831	       415.8 ns/op	     504 B/op	      18 allocs/op
BenchmarkMultiplyByPrice-8   	 2942138	       442.4 ns/op	     504 B/op	      18 allocs/op

After:
BenchmarkMultiplyByPrice-8   	 6882078	       209.4 ns/op	     232 B/op	       9 allocs/op
BenchmarkMultiplyByPrice-8   	 6852960	       174.1 ns/op	     232 B/op	       9 allocs/op
BenchmarkMultiplyByPrice-8   	 6674035	       175.2 ns/op	     232 B/op	       9 allocs/op
BenchmarkMultiplyByPrice-8   	 6824532	       178.8 ns/op	     232 B/op	       9 allocs/op
BenchmarkMultiplyByPrice-8   	 6637941	       176.8 ns/op	     232 B/op	       9 allocs/op
```

### Author/Reviewer Checklist
- [ ] If this PR has changes that result in a different app state given the same prior state and transaction list, manually add the `state-breaking` label.
- [ ] If the PR has breaking postgres changes to the indexer add the `indexer-postgres-breaking` label.
- [ ] If this PR isn't state-breaking but has changes that modify behavior in `PrepareProposal` or `ProcessProposal`, manually add the label `proposal-breaking`.
- [ ] If this PR is one of many that implement a specific feature, manually label them all `feature:[feature-name]`.
- [ ] If you wish to for mergify-bot to automatically create a PR to backport your change to a release branch, manually add the label `backport/[branch-name]`.
- [ ] Manually add any of the following labels: `refactor`, `chore`, `bug`.